Create a dedicated class for statistics
